新增 手机狂欢城脚本，互助码环境变量为 JD818_SHARECODES ， 互助码每天都变化
