+++ docker/auto_help.sh
@@ -103,7 +103,7 @@ autoHelp(){
 ############# 短期活动 #############
 
 #环球挑战赛
-autoHelp "环球挑战赛好友互助码" "${logDir}/jd_global.log" "JDGLOBAL_SHARECODES"
+autoHelp "京东手机狂欢城好友互助码" "${logDir}/jd_carnivalcity.log" "JD818_SHARECODES"
